Get rid of most truncation and unsigned-signed comparisons warnings also fixed mistake in previous commit

+++ wkbre2/NNSearch.cpp
@@ -27,7 +27,7 @@ PrgGameObject* NNSearch<Program, PrgGameObject>::next()
 {
 	while (true) {
 		const auto& vec = server->tiles[tz * server->terrain->getNumPlayableTiles().first + tx].objList;
-		if (it < vec.size()) {
+		if ((size_t)it < vec.size()) {
 			PrgGameObject* obj = vec[it++].getFrom<Program>();
 			if (obj)
 				return obj;
